+++ tests/compilation_failures/core/error_no_fields.rs
@@ -3,7 +3,7 @@ use preinterpret::*;
 macro_rules! assert_literals_eq {
     ($input1:literal, $input2:literal) => {stream!{
         [!if! ($input1 != $input2) {
-            #(%[].error(%["Expected " $input1 " to equal " $input2].string()))
+            #(%[].error(%["Expected " $input1 " to equal " $input2].to_string()))
         }]
     }};
 }