+++ tests/compilation_failures/core/error_span_single.rs
@@ -3,7 +3,7 @@ use preinterpret::*;
 macro_rules! assert_is_100 {
     ($input:literal) => {stream!{
         [!if! ($input != 100) {
-            #(%[$input].error(%["Expected 100, got " $input].string()))
+            #(%[$input].error(%["Expected 100, got " $input].to_string()))
         }]
     }};
 }