+++ tests/expressions.rs
@@ -44,7 +44,7 @@ fn test_basic_evaluate_works() {
     preinterpret_assert_eq!(#(let six_as_sum = 3 + 3; six_as_sum * six_as_sum), 36);
     preinterpret_assert_eq!(#(
         let partial_sum = %[+ 2];
-        %[#(%[5] + partial_sum) %[=] %raw[#](5 #partial_sum)].reinterpret_as_stream().debug_string()
+        %[#(%[5] + partial_sum) %[=] %raw[#](5 #partial_sum)].reinterpret_as_stream().to_debug_string()
     ), "%[5 + 2 = 7]");
     preinterpret_assert_eq!(#(1 + (1..2) as int), 2);
     preinterpret_assert_eq!(#("hello" == "world"), false);
@@ -56,19 +56,19 @@ fn test_basic_evaluate_works() {
     preinterpret_assert_eq!(#('A' < 'B'), true);
     preinterpret_assert_eq!(#("Zoo" > "Aardvark"), true);
     preinterpret_assert_eq!(
-        #(("Hello" as stream + "World" as stream + (1 + 1) as stream + (1 + 1).group() + [1 + 2].group() + %group[1 + 2 + 3]).debug_string()),
+        #(("Hello" as stream + "World" as stream + (1 + 1) as stream + (1 + 1).to_group() + [1 + 2].to_group() + %group[1 + 2 + 3]).to_debug_string()),
         r#"%["Hello" "World" 2 %group[2] %group[3] %group[1 + 2 + 3]]"#
     );
     preinterpret_assert_eq!(
-        #([1, 2, 1 + 2, 4].debug_string()),
+        #([1, 2, 1 + 2, 4].to_debug_string()),
         "[1, 2, 3, 4]"
     );
     preinterpret_assert_eq!(
-        #(([1 + (3 + 4), [5,] + [], [[6, 7],]] + [123]).debug_string()),
+        #(([1 + (3 + 4), [5,] + [], [[6, 7],]] + [123]).to_debug_string()),
         "[8, [5], [[6, 7]], 123]"
     );
     preinterpret_assert_eq!(
-        #(("Hello" as stream + "World" as stream + (1 + 1) as stream + (1 + 1).group()).debug_string()),
+        #(("Hello" as stream + "World" as stream + (1 + 1) as stream + (1 + 1).to_group()).to_debug_string()),
         r#"%["Hello" "World" 2 %group[2]]"#
     );
     assert_eq!(run!(let x = 1; x + 2), 3);
@@ -207,7 +207,7 @@ fn assign_works() {
     preinterpret_assert_eq!(
         #(
             let x = 5 + 5;
-            x.debug_string()
+            x.to_debug_string()
         ),
         "10"
     );
@@ -269,18 +269,18 @@ fn test_range() {
     );
     preinterpret_assert_eq!({ [!string! #(('a'..='f') as stream)] }, "abcdef");
     preinterpret_assert_eq!(
-        #((-1i8..3i8).debug_string()),
+        #((-1i8..3i8).to_debug_string()),
         "-1i8..3i8"
     );
 
     // Large ranges are allowed, but are subject to limits at iteration time
-    preinterpret_assert_eq!(#((0..10000).debug_string()), "0..10000");
-    preinterpret_assert_eq!(#((..5 + 5).debug_string()), "..10");
-    preinterpret_assert_eq!(#((..=9).debug_string()), "..=9");
-    preinterpret_assert_eq!(#((..).debug_string()), "..");
-    preinterpret_assert_eq!(#([.., ..].debug_string()), "[.., ..]");
-    preinterpret_assert_eq!(#([..[1, 2..], ..].debug_string()), "[..[1, 2..], ..]");
-    preinterpret_assert_eq!(#((4 + 7..=10).debug_string()), "11..=10");
+    preinterpret_assert_eq!(#((0..10000).to_debug_string()), "0..10000");
+    preinterpret_assert_eq!(#((..5 + 5).to_debug_string()), "..10");
+    preinterpret_assert_eq!(#((..=9).to_debug_string()), "..=9");
+    preinterpret_assert_eq!(#((..).to_debug_string()), "..");
+    preinterpret_assert_eq!(#([.., ..].to_debug_string()), "[.., ..]");
+    preinterpret_assert_eq!(#([..[1, 2..], ..].to_debug_string()), "[..[1, 2..], ..]");
+    preinterpret_assert_eq!(#((4 + 7..=10).to_debug_string()), "11..=10");
     preinterpret_assert_eq!(
         [!for! i in 0..10000000 {
             [!if! i == 5 {
@@ -290,7 +290,7 @@ fn test_range() {
         }],
         "5"
     );
-    // preinterpret_assert_eq!(#((0..10000 as iterator).debug_string()), "[<iterator> 0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, ..<9980 further items>]");
+    // preinterpret_assert_eq!(#((0..10000 as iterator).to_debug_string()), "[<iterator> 0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, ..<9980 further items>]");
 }
 
 #[test]
@@ -307,7 +307,7 @@ fn test_array_indexing() {
             let x = [0, 0, 0];
             x[0] = 2;
             (x[1 + 1]) = x[0];
-            x.debug_string()
+            x.to_debug_string()
         ),
         "[2, 0, 2]"
     );
@@ -315,42 +315,42 @@ fn test_array_indexing() {
     preinterpret_assert_eq!(
         #(
             let x = [1, 2, 3, 4, 5];
-            x.take()[..].debug_string()
+            x.take()[..].to_debug_string()
         ),
         "[1, 2, 3, 4, 5]"
     );
     preinterpret_assert_eq!(
         #(
             let x = [1, 2, 3, 4, 5];
-            x.take()[0..0].debug_string()
+            x.take()[0..0].to_debug_string()
         ),
         "[]"
     );
     preinterpret_assert_eq!(
         #(
             let x = [1, 2, 3, 4, 5];
-            x.take()[2..=2].debug_string()
+            x.take()[2..=2].to_debug_string()
         ),
         "[3]"
     );
     preinterpret_assert_eq!(
         #(
             let x = [1, 2, 3, 4, 5];
-            x.take()[..=2].debug_string()
+            x.take()[..=2].to_debug_string()
         ),
         "[1, 2, 3]"
     );
     preinterpret_assert_eq!(
         #(
             let x = [1, 2, 3, 4, 5];
-            x.take()[..4].debug_string()
+            x.take()[..4].to_debug_string()
         ),
         "[1, 2, 3, 4]"
     );
     preinterpret_assert_eq!(
         #(
             let x = [1, 2, 3, 4, 5];
-            x.take()[2..].debug_string()
+            x.take()[2..].to_debug_string()
         ),
         "[3, 4, 5]"
     );
@@ -364,7 +364,7 @@ fn test_array_place_destructurings() {
             let a = 0; let b = 0; let c = 0;
             let x = [1, 2, 3, 4, 5];
             [a, b, _, _, c] = x.take();
-            [a, b, c].debug_string()
+            [a, b, c].to_debug_string()
         ),
         "[1, 2, 5]"
     );
@@ -373,7 +373,7 @@ fn test_array_place_destructurings() {
             let a = 0; let b = 0; let c = 0;
             let x = [1, 2, 3, 4, 5];
             [a, b, c, ..] = x.take();
-            [a, b, c].debug_string()
+            [a, b, c].to_debug_string()
         ),
         "[1, 2, 3]"
     );
@@ -382,7 +382,7 @@ fn test_array_place_destructurings() {
             let a = 0; let b = 0; let c = 0;
             let x = [1, 2, 3, 4, 5];
             [.., a, b] = x.take();
-            [a, b, c].debug_string()
+            [a, b, c].to_debug_string()
         ),
         "[4, 5, 0]"
     );
@@ -391,7 +391,7 @@ fn test_array_place_destructurings() {
             let a = 0; let b = 0; let c = 0;
             let x = [1, 2, 3, 4, 5];
             [a, .., b, c] = x.take();
-            [a, b, c].debug_string()
+            [a, b, c].to_debug_string()
         ),
         "[1, 4, 5]"
     );
@@ -401,7 +401,7 @@ fn test_array_place_destructurings() {
             let out = [[0, 0], 0];
             let a = 0; let b = 0; let c = 0;
             [out[1], .., out[0][0], out[0][1]] = [1, 2, 3, 4, 5];
-            out.debug_string()
+            out.to_debug_string()
         ),
         "[[4, 5], 1]"
     );
@@ -415,7 +415,7 @@ fn test_array_place_destructurings() {
             let _ = c = [a[2], _] = [4, 5];
             let _ = a[1] += 2;
             let _ = b = 2;
-            [a.take(), b, c].debug_string()
+            [a.take(), b, c].to_debug_string()
         ),
         "[[0, 2, 4, 0, 0], 2, None]"
     );
@@ -426,7 +426,7 @@ fn test_array_place_destructurings() {
             let a = [0, 0];
             let b = 0;
             a[b] += #(b += 1; 5);
-            a.debug_string()
+            a.to_debug_string()
         ),
         "[0, 5]"
     );
@@ -451,35 +451,35 @@ fn test_array_pattern_destructurings() {
     preinterpret_assert_eq!(
         #(
             let [a, b, _, _, c] = [1, 2, 3, 4, 5];
-            [a, b, c].debug_string()
+            [a, b, c].to_debug_string()
         ),
         "[1, 2, 5]"
     );
     preinterpret_assert_eq!(
         #(
             let [a, b, c, ..] = [1, 2, 3, 4, 5];
-            [a, b, c].debug_string()
+            [a, b, c].to_debug_string()
         ),
         "[1, 2, 3]"
     );
     preinterpret_assert_eq!(
         #(
             let [.., a, b] = [1, 2, 3, 4, 5];
-            [a, b].debug_string()
+            [a, b].to_debug_string()
         ),
         "[4, 5]"
     );
     preinterpret_assert_eq!(
         #(
             let [a, .., b, c] = [1, 2, 3, 4, 5];
-            [a, b, c].debug_string()
+            [a, b, c].to_debug_string()
         ),
         "[1, 4, 5]"
     );
     preinterpret_assert_eq!(
         #(
             let [a, .., b, c] = [[1, "a"], 2, 3, 4, 5];
-            [a.take(), b, c].debug_string()
+            [a.take(), b, c].to_debug_string()
         ),
         r#"[[1, "a"], 4, 5]"#
     );
@@ -495,25 +495,25 @@ fn test_objects() {
             x["x y z"] = 4;
             x["z\" test"] = %{};
             x.y = 5;
-            x.debug_string()
+            x.to_debug_string()
         ),
         r#"{ a: {}, b: "Hello", hello: 1, world: 2, ["x y z"]: 4, y: 5, ["z\" test"]: {} }"#
     );
     preinterpret_assert_eq!(
         #(
-            %{ prop1: 1 }["prop1"].debug_string()
+            %{ prop1: 1 }["prop1"].to_debug_string()
         ),
         r#"1"#
     );
     preinterpret_assert_eq!(
         #(
-            %{ prop1: 1 }["prop2"].debug_string()
+            %{ prop1: 1 }["prop2"].to_debug_string()
         ),
         r#"None"#
     );
     preinterpret_assert_eq!(
         #(
-            %{ prop1: 1 }.prop1.debug_string()
+            %{ prop1: 1 }.prop1.to_debug_string()
         ),
         r#"1"#
     );
@@ -523,14 +523,14 @@ fn test_objects() {
             let b;
             let z;
             %{ a, y: [_, b], z } = %{ a: 1, y: [5, 7] };
-            %{ a, b, z }.debug_string()
+            %{ a, b, z }.to_debug_string()
         ),
         r#"{ a: 1, b: 7, z: None }"#
     );
     preinterpret_assert_eq!(
         #(
             let %{ a, y: [_, b], ["c"]: c, [r#"two "words"#]: x, z } = %{ a: 1, y: [5, 7], ["two \"words"]: %{}, };
-            %{ a, b, c, x: x.take(), z }.debug_string()
+            %{ a, b, c, x: x.take(), z }.to_debug_string()
         ),
         r#"{ a: 1, b: 7, c: None, x: {}, z: None }"#
     );
@@ -550,18 +550,18 @@ fn test_method_calls() {
             let x = [1, 2, 3];
             x.push(5);
             x.push(2);
-            x.debug_string()
+            x.to_debug_string()
         ),
         "[1, 2, 3, 5, 2]"
     );
     // Push returns None
     preinterpret_assert_eq!(
-        #([1, 2, 3].as_mut().push(4).debug_string()),
+        #([1, 2, 3].as_mut().push(4).to_debug_string()),
         "None"
     );
     // Converting to mut and then to shared works
     preinterpret_assert_eq!(
-        #([].as_mut().len().debug_string()),
+        #([].as_mut().len().to_debug_string()),
         "0usize"
     );
     preinterpret_assert_eq!(
@@ -569,7 +569,7 @@ fn test_method_calls() {
             let x = [1, 2, 3];
             let y = x.take();
             // x is now None
-            x.debug_string() + " - " + y.debug_string()
+            x.to_debug_string() + " - " + y.to_debug_string()
         ),
         "None - [1, 2, 3]"
     );
@@ -592,7 +592,7 @@ fn stream_append_can_use_self_in_appender() {
         run! {
             let variable = %[Hello];
             variable += %[World #variable];
-            variable.debug_string()
+            variable.to_debug_string()
         },
         "%[Hello World Hello]"
     );
@@ -600,7 +600,7 @@ fn stream_append_can_use_self_in_appender() {
         run! {
             let variable = %[Hello];
             variable += %[World #(variable += %[!];)];
-            variable.debug_string()
+            variable.to_debug_string()
         },
         "%[Hello ! World]"
     );
@@ -608,7 +608,7 @@ fn stream_append_can_use_self_in_appender() {
         run! {
             let variable = %[Hello];
             variable += %[World #(variable = %[Hello2];)];
-            variable.debug_string()
+            variable.to_debug_string()
         },
         "%[Hello2 World]"
     );